implement search

+++ app/bootstraps/Bootstraper.java
@@ -5,21 +5,23 @@ import java.util.List;
 import play.jobs.*;
 import models.*;
 import play.test.Fixtures;
- 
+
 @OnApplicationStart
 public class Bootstraper extends Job {
-    
-    public void doJob() {
-      Fixtures.deleteAll();
-      /*Fixtures.load("jobs/bootstraps/diploma.yml");*/
-      Fixtures.load("bootstraps/user.yml");
-      List<User> users = User.findAll();
-      for(User u : users)
-          for(User user : users)
-              if(!u.equals(user)) 
-                  user.friends.add(u); // make users all friends 
-      for(User u : users)
-          u.save();
-    }
-    
+
+	public void doJob() {
+		if(User.count()==0) {
+			Fixtures.deleteAll();
+			/* Fixtures.load("jobs/bootstraps/diploma.yml"); */
+			Fixtures.load("bootstraps/user.yml");
+			List<User> users = User.findAll();
+			for (User u : users)
+				for (User user : users)
+					if (!u.equals(user))
+						user.friends.add(u); // make users all friends
+			for (User u : users)
+				u.save();
+		}
+	}
+
 }
